% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clogistic1.R
\name{Clogistic1}
\alias{Clogistic1}
\title{Clustered coefficient logistic regression model with common effects}
\usage{
Clogistic1(
  indexy,
  y,
  z,
  x,
  sweights,
  Ni,
  betam0,
  eta00,
  nu = 1,
  gam = 3,
  lam = 0.5,
  inner_loop = 100,
  maxiter = 1000,
  tol = 0.001,
  tol1 = 0.001,
  tol2 = 0.001
)
}
\arguments{
\item{indexy}{index for y (area), should be same for same areas or units}

\item{y}{response}

\item{z}{covariates with common effects}

\item{x}{covariates with clustered effects}

\item{sweights}{sampling weights}

\item{Ni}{population size in each unit (area); a vector of m (number of units)}

\item{betam0}{initial values for coefficient of x}

\item{eta00}{initial values for z}

\item{nu}{a fixed value in ADMM}

\item{gam}{a fixed value in SCAD}

\item{lam}{tuning parameter}

\item{tol2}{}
}
\description{
Clustered coefficient logistic regression model with common effects
}
