% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clogistic_bic.R
\name{Clogistic_bic}
\alias{Clogistic_bic}
\title{Get clustered estimates based on given data, model and lambda search}
\usage{
Clogistic_bic(
  area,
  y,
  x,
  wts,
  N,
  model,
  group = NULL,
  lambda,
  standardize = TRUE,
  center = TRUE
)
}
\arguments{
\item{area}{area vector}

\item{y}{response, sampled data}

\item{x}{covariates without intercept, sampled data}

\item{wts}{sampling weights}

\item{N}{population size in each unit (area); a vector of m (number of units)}

\item{model}{three models, "intercept","creg" and "ccreg"}

\item{group}{group information for ccreg model}

\item{lambda}{a vector of lambda values}
}
\description{
Get clustered estimates based on given data, model and lambda search
}
