% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clogistic_group.R
\name{Clogistic_group}
\alias{Clogistic_group}
\title{Clustered coefficient logistic regression model with group covariates}
\usage{
Clogistic_group(
  indexy,
  y,
  x,
  sweights,
  Ni,
  group,
  betam0,
  nu = 1,
  gam = 3,
  lam = 0.5,
  inner_loop = 500,
  maxiter = 1000,
  tol = 0.001,
  tol1 = 0.001,
  tol2 = 0.001
)
}
\arguments{
\item{indexy}{index for y (area), should be same for same areas or units}

\item{y}{response}

\item{x}{covariates with clustered effects}

\item{sweights}{sampling weights}

\item{Ni}{population size in each unit (area); a vector of m (number of units)}

\item{group}{group index for covariates}

\item{betam0}{initial values for coefficient of x}

\item{nu}{a fixed value in ADMM}

\item{gam}{a fixed value in SCAD}

\item{lam}{tuning parameter}

\item{tol2}{}

\item{z}{covariates with common effects}

\item{eta00}{initial values for z}
}
\description{
Clustered coefficient logistic regression model with group covariates
}
