% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clogistic_SAE.R
\name{Clogistic_SAE}
\alias{Clogistic_SAE}
\title{Get estimates  based on given data, model and lambda search}
\usage{
Clogistic_SAE(area, y, x, xunit, wts, N, model, group = NULL, lambda)
}
\arguments{
\item{area}{area vector}

\item{y}{response, sampled data}

\item{x}{covariates without intercept, sampled data}

\item{xunit}{population level data, first column the area index}

\item{wts}{sampling weights}

\item{N}{population size in each unit (area); a vector of m (number of units)}

\item{model}{three models, "intercept","creg" and "ccreg"}

\item{group}{group information for ccreg model}

\item{lambda}{a vector of lambda values}
}
\description{
Get estimates  based on given data, model and lambda search
}
