% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estSAE_logistic.R
\name{estSAE_intercept}
\alias{estSAE_intercept}
\title{estimation for area proportions based on clustered intercept model}
\usage{
estSAE_intercept(obj, indexy, y, z, sweights, zpopd, Ni)
}
\arguments{
\item{obj}{model object}

\item{indexy}{area index}

\item{y}{response}

\item{z}{covariates common regression coef}

\item{sweights}{sampling weights}

\item{zpopd}{In each area, the number of units for each combination,
first column area, second column number of units in each domain Nc
same as original z in a dummies form with same column names}

\item{Ni}{population size in each unit (area); a vector of m (number of units)}
}
\description{
estimation for area proportions based on clustered intercept model
}
