% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estSAE_logistic.R
\name{estSAE_logistic}
\alias{estSAE_logistic}
\title{estimation for area proportions based on models three estimator}
\usage{
estSAE_logistic(
  obj,
  indexy,
  y,
  x,
  xunit,
  wts,
  N,
  model,
  group = NULL,
  index_d = NULL
)
}
\arguments{
\item{obj}{model object}

\item{indexy}{area index}

\item{y}{response}

\item{x}{covariates}

\item{xunit}{population level x}

\item{N}{population size in each unit (area); a vector of m (number of units)}

\item{sweights}{sampling weights}
}
\description{
estimation for area proportions based on models three estimator
}
