% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial1.R
\name{initial1}
\alias{initial1}
\title{initial values for Clogistic1}
\usage{
initial1(
  indexy,
  y,
  z,
  x,
  sweights,
  Ni,
  lam0 = 1e-04,
  tol = 1e-05,
  maxiter = 500
)
}
\arguments{
\item{indexy}{index for y (area), should be same for same areas or units}

\item{y}{response}

\item{z}{covariates with common effects}

\item{x}{covariates with clustered effects}

\item{sweights}{sampling weights}

\item{Ni}{population size in each unit (area); a vector of m (number of units)}

\item{maxiter}{}
}
\description{
initial values for Clogistic1
}
