% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial2.R
\name{initial2}
\alias{initial2}
\title{initial values for Clogistic2 and Clogistic_group}
\usage{
initial2(indexy, y, x, sweights, Ni, lam0 = 0.01, tol = 1e-05, maxiter = 500)
}
\arguments{
\item{indexy}{index for y (area), should be same for same areas or units}

\item{y}{response}

\item{x}{covariates with clustered effects}

\item{sweights}{sampling weights}

\item{Ni}{population size in each unit (area); a vector of m (number of units)}

\item{maxiter}{}
}
\description{
initial values for Clogistic2 and Clogistic_group
}
