% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgreg.R
\name{lgreg}
\alias{lgreg}
\title{estimation for area proportions based on model with common regression coefficients}
\usage{
lgreg(indexy, y, x, sweights, xunit, N)
}
\arguments{
\item{indexy}{area index}

\item{y}{response}

\item{x}{covariates common regression coef}

\item{sweights}{sampling weights}

\item{xunit}{In each area, population level}

\item{N}{population size in each unit (area); a vector of m (number of units)}

\item{obj}{model object}
}
\description{
estimation for area proportions based on model with common regression coefficients
}
