% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_logistic.R
\name{refit_Lm1}
\alias{refit_Lm1}
\title{refit based on given cluster information and sampling weights three models}
\usage{
refit_Lm1(indexy, y, z, x, cluster, wts, N)
}
\arguments{
\item{indexy}{area index}

\item{y}{response}

\item{cluster}{cluster structure}

\item{wts}{sampling weights}

\item{N}{population size in each unit (area); a vector of m (number of units)}

\item{obj}{model object}
}
\description{
refit based on given cluster information and sampling weights three models
}
